requirements.txt
opencv-python
opensimplex~=0.4.5.1
Pillow~=11.3.0
numpy~=2.3.3

numba~=0.62.0
scipy~=1.16.2

PySide6~=6.9.2
NodeGraphQt~=0.6.42
vispy~=0.15.2
setuptools


psutil~=7.1.0

ruff
Mypy
types-psutil

landlab

PyOpenGL